day64/top-movies/requirements.txt
Bootstrap_Flask==2.3.0
Flask==3.0.0
flask_sqlalchemy==3.1.1
Flask_WTF==1.2.1
Requests==2.31.0
WTForms==3.0.1
python-dotenv==1.0.0